Move server config to Configuration.kt
