+++ server/src/main/kotlin/org/javacs/kt/KotlinLanguageServer.kt
@@ -7,8 +7,6 @@ import org.eclipse.lsp4j.services.LanguageClient
 import org.eclipse.lsp4j.services.LanguageClientAware
 import org.eclipse.lsp4j.services.LanguageServer
 import org.javacs.kt.command.ALL_COMMANDS
-import org.javacs.kt.config.ServerConfiguration
-import org.javacs.kt.config.parseServerConfiguration
 import org.javacs.kt.externalsources.ClassPathSourceJarProvider
 import org.javacs.kt.externalsources.JarClassContentProvider
 import org.javacs.kt.progress.LanguageClientProgress
@@ -43,8 +41,6 @@ class KotlinLanguageServer : LanguageServer, LanguageClientAware, Closeable {
             field = factory
             sourcePath.progressFactory = factory
         }
-
-    private var serverConfiguration: ServerConfiguration? = null
 
     companion object {
         val VERSION: String? = System.getProperty("kotlinLanguageServer.version")
@@ -92,9 +88,9 @@ class KotlinLanguageServer : LanguageServer, LanguageClientAware, Closeable {
         serverCapabilities.documentRangeFormattingProvider = Either.forLeft(true)
         serverCapabilities.executeCommandProvider = ExecuteCommandOptions(ALL_COMMANDS)
 
-        serverConfiguration = parseServerConfiguration(params)
+        config.server = parseServerConfiguration(params)
 
-        classPath.storage = serverConfiguration?.storage?.getSlice("classpath")
+        classPath.storage = config.server?.storage?.getSlice("classpath")
 
         val clientCapabilities = params.capabilities
         config.completion.snippets.enabled = clientCapabilities?.textDocument?.completion?.completionItem?.snippetSupport ?: false
